+++ .gitignore
@@ -1,4 +1,3 @@
 node_modules
 /photos/
-/database/mockData.json
-/database/mockData.csv
+/database/datasets/
