--- database/postgres/test.csv
@@ -1,2 +1,0 @@
-city,title,hostImage,roomInfo,numberOfGuests,numberOfBedrooms,numberOfBeds,numberOfBaths,isSuperhost,isGreatLocation,isSparklingClean,isGreatCheckIn,isSelfCheckIn,description,amenities.basic.hasWiFi,amenities.basic.hasEssentials,amenities.basic.hasCable,amenities.basic.hasLaptopSpace,amenities.basic.hasHeating,amenities.dining.hasKitchen,amenities.bedAndBath.hasPillowsBlankets,sleepingArrangements.bedroom,id
-Lake Alessandro,Stroll Around Victoria Park Lake Alessandro,https://sdc-airdnd.s3-us-west-2.amazonaws.com/host8.jpg,Entire guesthouse,6,3,4,2,false,false,true,true,true,Optio ut aut eum. Porro aut sint consectetur minus aut. Non ipsum laborum sunt totam labore nihil non. Fuga veniam eius veniam. Libero quia dolorem id aspernatur quae. Enim earum officia odit consequatur quasi qui qui error in.Et commodi et et hic ipsam voluptatum. Aut ducimus pariatur autem rem ab quasi. Blanditiis assumenda dolor ut possimus aspernatur perferendis. Reiciendis soluta eum.,true,true,true,true,true,true,true,3,1
