+++ package.json
@@ -43,6 +43,7 @@
     "webpack-cli": "^3.3.2"
   },
   "dependencies": {
+    "arangojs": "^7.0.1",
     "axios": "^0.20.0",
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
@@ -53,6 +54,7 @@
     "jsonexport": "^3.0.1",
     "mongoose": "^5.5.9",
     "nodemon": "^1.19.0",
+    "pg": "^8.3.3",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-modal": "^3.8.1",
